Fix data type of UTF32 string on 64-bit systems
long is 64-bit long on some LP64 systems

git-svn-id: https://s1.saphirion.com/svn/rm-asset/trunk/r3@5442 0b15d6fc-ce89-4483-8625-0b0fabc32e25

+++ src/os/host-text.c
@@ -80,7 +80,7 @@ static u32* text_ext_words;
 {
 	int res, len, n;
 	void *str;
-	unsigned long *wstr;
+	u32 *wstr;
 
 	if ((res = RL_Get_String(series, 0, &str)) == 0) {
 		*string = NULL;
@@ -88,7 +88,7 @@ static u32* text_ext_words;
 	}
 	
 	len = abs(res);
-	wstr = (unsigned long *)OS_Make((len+1) * sizeof(unsigned long));
+	wstr = (u32 *)OS_Make((len+1) * sizeof(u32));
 	
 	if (res < 0) {
 		for (n = 0; n < len; n++)
